feat: move stage list command to `state list`
