+++ cmd/sst/main.go
@@ -811,7 +811,6 @@ var root = &cli.Command{
 			},
 			Run: CmdShell,
 		},
-		CmdListStages,
 		{
 			Name: "remove",
 			Description: cli.Description{
